+++ internal/controller/pipelinerollout_controller.go
@@ -410,6 +410,10 @@ func (r *PipelineRolloutReconciler) processExistingPipeline(ctx context.Context,
 		return fmt.Errorf("failed to convert existing Pipeline spec %q into PipelineSpec type, err=%v", string(existingPipelineDef.Spec.Raw), err)
 	}
 
+	// TTODO:
+	// usde.GetUpgradeStrategy(newPipelineDef, existingPipelineDef) => do-nothing, direct-apply, ppnd?
+	// based on function returned value, perform appropriate operations below and/or pass the value around
+	// INSTEAD OF:
 	// Does pipeline spec need to be updated?
 	pipelineNeedsToUpdate, err := pipelineNeedsUpdating(ctx, newPipelineDef, existingPipelineDef)
 	if err != nil {
